# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Закрытый API для системы управления перевозками",
        "description": "API documentation for My Application",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "andrew.izmaylov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://transportation.test/public/api/v1/",
            "description": "Local development server"
        },
        {
            "url": "https://transportation.test/public/api/v1/",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/shipper/{transportation_id}/{cargo_id}/delete-cargo": {
            "delete": {
                "tags": [
                    "4. Shipper - Карго"
                ],
                "summary": "Удаление карго",
                "description": "Удаляет карго из доставки. Идентификатор перевозки и карго передается в параметрах URI",
                "operationId": "d9c5f270693cd556bc449ae950d6a36a",
                "parameters": [
                    {
                        "name": "transportation_id",
                        "in": "path",
                        "description": "Transportation ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address_id",
                        "in": "path",
                        "description": "Address ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TransportationAddress deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransportationCard"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/shipper/{transportation_id}/{cargo_id}/update-cargo": {
            "patch": {
                "tags": [
                    "4. Shipper - Карго"
                ],
                "summary": "Изменение груза",
                "description": "Изменяет данные груза. Идентификатор перевозки и адреса передается в параметрах URI",
                "operationId": "18f669e76247d289dcaef6cc1080bedb",
                "parameters": [
                    {
                        "name": "transportation_id",
                        "in": "path",
                        "description": "Transportation ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cargo_id",
                        "in": "path",
                        "description": "Cargo ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления объекта карго",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Чемодан"
                                    },
                                    "length": {
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "width": {
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "height": {
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "weight": {
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "65000"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "RUB"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cargo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransportationCard"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "TransportationCard": {
                "required": [
                    "id",
                    "type",
                    "attributes"
                ],
                "properties": {
                    "id": {
                        "description": "Transportation ID",
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "example": "Transportation"
                    },
                    "attributes": {
                        "required": [
                            "id",
                            "name",
                            "clientId",
                            "pickupFrom",
                            "pickupTo",
                            "transportationStatus"
                        ],
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "clientId": {
                                "type": "string"
                            },
                            "pickupFrom": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-01 10:00:00"
                            },
                            "pickupTo": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-01 18:00:00"
                            },
                            "transportationStatus": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}